feat(mem)!: rewrite the memory module
Before this commit, a continuous memory region can only be backed by
either a single ArcMemPages or a single MmioRange.
If a guest memory region contains both mmap()-ed pages and emulated
MMIO regions, we then need to split it into sub-regions. However,
once these sub-regions are added to the memory bus as slots, we can
no longer tell if a slot is a sub-region or a not.

It becomes difficult to handle when a guest tries to relocate
a region to a different physical address. For example, a PCI device
can re-program its BARs. We do not know if a few slots need to be
moved as a whole or not.

This commit re-designs the module `mem` by introducing `MemRegion`.
* From guest's point of view, it can have multiple sub-regions of
  different e820 types.
* From host's point of view, it is a collection of mmap()-ed host
  memories, emulated MMIO regions, and maybe some holes in between.
  Backends are grouped so they can be moved easily.

It also enables us to allocate the address for a group of slots.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>