style: use `foo/foo.rs` as module paths
Rust 2018 introduced a new directory structure `foo.rs + foo/bar.rs`
to replace the old `foo/mod.rs + foo/bar.rs` for a module `foo` with
submodule `bar`. While it solved the issue of "a bunch of tabs named
mod.rs", it introduced a new problem. If a lib has multiple modules
with submodules, most file browsers list directories before files,

```
├── lib.rs
├── foo1
│   └── bar1.rs
├── foo2
│   └── bar2.rs
├── foo3
│   └── bar3.rs
├── foo1.rs
├── foo2.rs
└── foo3.rs
```

and thus visually `foo1` is far from `foo1.rs`.

This commit introduces a customized file hierarchy: for a module
`foo` and its submodule `bar`, we place the source code in
`foo/foo.rs` and `foo/bar.rs` respectively, so we can avoid `mod.rs`
and also put source files of a module in a single directory.

For sure this breaks the community convention and we have to use the
`path` attribute every time.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>