fix(virtio)!: enable ACCESS_PLATFORM for CoCo only
The bit VIRTIO_F_ACCESS_PLATFORM is a little bit confusing:

- when a Linux guest driver sees this bit, it uses DMA API for virtq
  transactions.
  - For a confidential VM, the DMA layer will use swiotlb to copy
    data from/to the shared memory.
  - In the non-CoCo case, I observed 5% performance regression of the
    virtio-net throughput.
- when an in-kernel vhost device sees this bit, addresses in the
  virtq are viewed as io virtual addresses and it expects the
  userspace VMM to setup the IOTLB.

We do not have an emulated IOMMU, so we should not set
ACCESS_PLATFORM on backend devices. On the other hand, in the CoCo
case, we must advertise this bit to the guest.

So for now, the solution is, we let the middle layer turn on
ACCESS_PLATFORM if necessary but never activate the device backend
with this bit. We will need to find a better solution in the future.

Fixes: 5ad2ea658cf2 ("feat(virtio): enable ACCESS_PLATFORM bits")

Signed-off-by: Changyuan Lyu <changyuanl@google.com>