feat(pci)!: reserve a BDF before adding a device
MSI on aarch64 requires an additional devid, which is usually the
device BDF. Thus the `MsiSender` of a PCI device needs to know the
BDF. However, currently the BDF is unknown until the device
has been plugged into the PCI bus.

This commit changes the API such that a BDF is reserved first before
a device is plugged in. The reserved BDF then can be passed to the
constructor of an `MsiSender`.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>