fix(mem): do not add device memory to `RamBus`
This commit follows 04a9c4de56f6 ("fix(mem): differentiate device
memory from RAM").

`RamBus` is the DMA interface for devices. PCI memory resources and
the firmware should not be added to `RamBus`. One exception is that
virtio-mem share memory regions should be added to the `RamBus`. In
that case the constructor of a future virtio-mem device should return
a `MemRange::Ram` in its `shared_mem_regions()` method.

The control plane is moved out of `RamBus` such that device memory can
be mapped to the VM but not added to `RamBus`.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>