feat(cpu): introduce CPU topology configuration
This commit adds option `topology` to flag --cpu to allow users to
specify CPU topology.

On x86_64, the CPU topology is encoded in CPU x2APIC ID, as well CPUID
leaf 0xb, 0x1f (Intel Only), and 0x8000_0026 (AMD only). On aarch64, the
CPU topology is encoded in CPU MPIDR and the devicetree.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>