fix(dt): preserve device tree node order
Linux assigns indexes to CPUs based on the order of the corresponding
CPU nodes in the devicetree. Using a HashMap for sub-nodes does not
guarantee any order.

To ensure the flattened devicetree has the same sub-node order as the
CPU nodes are created, this commit changes the representation of
devicetree sub-nodes from a HashMap to a Vec of key-value pairs.

Signed-off-by: Changyuan Lyu <changyuanl@google.com>